Revert^2 "Turn on V3 cow"
This reverts commit a256b813f9197105a6984044973864ff904ce563.

Reason for revert: we were running into OTA failures b/333966507. The reason was that OTA compatibility tools on go/absign weren't updated with our v3 compatibility tools (which turns of v3 full OTA for devices launching prior to Android 15). Once we know that go/absign is updated, we can check in this change again

Change-Id: I789c24b57754d6ede794d7f9451ee0ca469c2fb4