Add soong installed files to general-tests.zip
The module `test_module_config` will be installing symlinks in the
testcases directory pointing back to the `base` test for .apk files.
This saves a gigs of space on disk and in suite zip files.

However, the .mk files were using  LOCAL_COMPATIBILITY_SUPPORT_FILE to
do two things:
  1) copy the file from intermediates to testcases
  2) package the testcases files in zips.

Because the test_module_config rule is taking care of 1), we need a new
variable to do just 2)

Test: m general-tests device-tests
Change-Id: Icdad4401cde27d8edd5e13e486b96c62817a4bbb