Add defaults target for shared linking
Add a defaults target that all targets linking to aconfig codegen with
the new storage API dependencies will need. The target takes the new
storage APIs conditionally behind a flag, so we can disable it easily.

In this topic, we add it to a single target to see that it builds and
adds the dependency. In follow-on CLs we can add it to the rest of the
targets.

Test: m audiopolicy_tests
Bug: 347701310
(cherry picked from https://android-review.googlesource.com/q/commit:71c3a1a365e37dcca9a1d5a8ac4d5c8a94b526ed)
Merged-In: I67643205b283301f72c5466401060212b637388f
Change-Id: I67643205b283301f72c5466401060212b637388f