Install dexpreopt artifacts of non system apps in system_other
`SYSTEM_OTHER_ODEX_FILTER` is used to determine which odex files go in
system_other partition. Soong and Make adds another filter on top of that.
This additional filter limits the odex files to only system apps, which
means that dexpreopt artifacts of product and system_ext apps are **not**
installed in system_other.

This CL removes this additional filter. After this CL, dexpreopt
artifacts of product and system_ext apps will be installed in system_other.

Bug: 349083274
Test: Verified that .odex files of product apps appear in
intalled-files-system_other.txt and not installed-files-product.txt

Test: Ran this CL on a bunch of targets in git_main and throttled
branchesa (there are some unrelated errors on some targets)
https://android-build.corp.google.com/builds/abtd/run/L93000030004826539
https://android-build.corp.google.com/builds/abtd/run/L53400030004824781
https://android-build.corp.google.com/builds/abtd/run/L41900030004824724
https://android-build.corp.google.com/builds/abtd/run/L87200030004822630
https://android-build.corp.google.com/builds/abtd/run/L67500030004727048
https://android-build.corp.google.com/builds/abtd/run/L41700030004726610
https://android-build.corp.google.com/builds/abtd/run/L61600030004726607

Change-Id: I522b8f00a268058b13bb88ac1b34824bd5c1190e