Add libmonkey_jni to base
The monkey package is already part of "base" build. In this topic, we
are adding jni support to monkey, which will allow us to use libinput's
uinput infrastructure for motion injection.

In this CL, we also need to add libmonkey_jni to the base build. Without
this change, monkey fails to find the jni library, because 'flashall'
does not include this .so by default.

However, this change is not needed when running 'adb sync', which
correctly recognizes that the jni library is required and places it in
the right place on the device.

Bug: 344695913
Flag: EXEMPT refactor
Ignore-AOSP-First: relies on uinput infra not available in aosp
Test: adb shell monkey 1000 --pct-touch 100
Test: m && vendor/google/tools/flashall -w --disable_verity=true
Change-Id: Id6c381895e914f0c0ab8278fad31e82734f48cf5