Drop dexpreopt files of service-compos from install diff allowlist
service-compos is an apex system server, and its parent apex
(com.android.compos) is installed in system_ext. The dexpreopt files of
all apex system server jars were previously installed in /system, but
with https://r.android.com/3288083, they will be installed in the same
partition as the parent apex.

Test: presbumits
Bug: 369678122
Change-Id: Ic32bcfd98667ad8bc92b7a747e702d18e7516ee9