Explicitly declare selinux_policy_(vendor|odm) in PRODUCT_PACKAGES
As part of the mk->bp migration, we are currently using heuristics to
classify the packages in `PRODUCT_PACKAGES` to deps of the different
android_filesystem modules. To aid such heuristics, separate phony
targets will be created for selinux_policy_vendor and
selinux_policy_odm. The former wil be added to deps of the autogenerated
vendor module and the latter will be added to deps of the autogenerated
odm module

Bug: 376755916
Test: Verified that NOTICE.xml.gz is the only diff in fileList contents
between kati and soong odm.img

Change-Id: I001080011e9a89b8053e91b042ed82e215b0187e