Add --allow-read-write flag to aconfig
This flag allows the user to specify whether or not to allow flags with READ_WRITE permission to be parsed. By default, the flag is set to true, which means that flags with READ_WRITE permission will be parsed. If the flag is set to false, then it is an error if flags with READ_WRITE permission are provided to the create-cache command.

Bug: 377294922
Test: cargo test

Change-Id: I48583a35e04d392fa7954d69e18884f2a7d46f35