Use --steps_for_build_test_only option to skip some steps
We don't need to run all finalization steps on each CI build target.
Use --steps_for_build_test_only option to skip ABI dump and VINTF
check steps for CI build tests. It is sufficient to build the full
steps on a single target but not all.
The finalization takes 30 minutes with this option. It took 2 hours
to complete full steps without the option.

Ignore-AOSP-First: finalization script is an exception
Bug: 382626911
Test: build/make/tools/finalization/build-step-0.sh \
          --steps_for_build_test_only && m
Change-Id: Ib4f04a2d2b4e54c4d9631be535ba0f44b1080fb1