Remove empty rules for hostname and build_number
These files are created by soong_ui and not by the primary ninja
execution. Empty rules are currently added to the ninja file to prevent
"Test for dangling rules" from flagging this as an error.

In preparation for turning on "Test for dangling rules" check in `m
--soong-only`, these files will be added to an allowlist in soong_ui's
test_build.go. Therefore, we can clean these up now.

Test: m checkbuild --soong-only
Test: presubmits
Bug: 394096939
Change-Id: I6fd3717fa52ff2461965c3d401860edfbc0cf57d