Don't read through symlinks in fsverity_metadata_generator
We're trying to change the build system to remove the need for a
staging directory, where all the files are in the same location
as they are on the device. This means that fsverity_metadata_generator
will no longer be able to read through symlinks, because the files
that the symlink is trying to point to won't be at the same location.

Instead, when seeing a symlink, make the corresponding fsv_meta file
also a symlink.

Bug: 394404628
Test: atest ComposHostTestCases
Change-Id: I857f135ff0825971b7779719576e2f931995e481