build: core: Add 16kb modules loaded in Second Stage to vendor_dlkm
This change will add the 16kb modules that are loaded in Second
Boot Stage to the vendor_dlkm partition.

The 16kb modules will be in the directory:

    /vendor_dlkm/lib/modules/16k-mode

This directory will only be available when the
16kb developer option is enabled.

Bug: 382569981
Test: Built and booted husky in 4k/16k modes
Flag: EXEMPT bug_fix
Ignore-AOSP-First: This change will be merged in AOSP once the whole top has been submitted.
Change-Id: Iaa3a0117184307da7de43cb7dd185b5f457c7bba