Add an aosp_arm64_plus_armv7 product for building test suites
CTS currently uses a hack in the aosp_arm64 BoardConfig.mk to switch
the 32-bit TARGET_2ND_ARCH_VARIANT to armv7 so that the 32-bit binaries
in android-cts.zip can be used on older devices.  That is probably
obsolete, but to maintain the current behavior while removing the
hack in aosp_arm64 to reduce reanalysis, introduce a separate product
for test suites builds to use.

Bug: 344958592
Bug: 389145684
Test: lunch aosp_arm64_plus_armv7 && get_build_var TARGET_2ND_ARCH_VARIANT
Change-Id: I1d697d24387d761cc8e45c2b1aac531e28289eb3