Add missing type parameters
Previously, when reading a signature file if Metalava came across an
unqualified type that is not a type parameter it would assume that it
was part of `java.lang` package. However, it now checks to see if that
is true and if not then it reports an error, which by default is thrown
as an exception. That breaks this test as it references `T` and `U`
which are clearly type parameters but they have never been defined.

This change fixes that by adding the missing type parameters.

Bug: 394789173
Test: atest check-flagged-apis-test
Change-Id: Ic5772c544f435c1a27d3ac0517b741a68670595c